pvoutput-publisher/requirements.txt
configobj==5.0.6
paho-mqtt==1.3.1
